[#61] IntoClickhouse trait을 mod.rs 파일로 이동
