+++ src/adapter/clickhouse.rs
@@ -1,6 +1,6 @@
 use serde::{Deserialize, Serialize};
 
-use crate::{adapter::convert::IntoClickhouseValue, errors};
+use crate::{adapter::IntoClickhouseValue, errors};
 
 #[derive(Clone)]
 pub struct ClickhouseConnection {
