+++ src/adapter/postgres/pgoutput.rs
@@ -3,7 +3,7 @@ use std::io::Read;
 use byteorder::ReadBytesExt;
 use serde::{Deserialize, Serialize};
 
-use crate::{adapter::convert::IntoClickhouseValue, errors};
+use crate::{adapter::IntoClickhouseValue, errors};
 
 #[derive(Debug, Clone, Copy, PartialEq, Serialize, Deserialize)]
 #[repr(u8)]
